Add databaseConfig and JDBC related classes
Databaseconfig reads configurations from the .config/database file, fetching the username, password and connectionstring neccessary to connect to the database.
The JDBCRepository is an abstract class that uses the dbConfig to connect to the database. By making it abstract each jdbcRepository can extend it, automatically gaining access to these methods.